Support for winrm-elevated version 0.4.0
winrm-elevated version 0.4.0 changes the initializer argument from
the previous winrm service object to the newer executor object.
Change miq_winrm to pass the executor.

This change now causes the WinRM-Elevated object to go out to the server
to get the os version where previously it simply returned a vanilla
object without any network traffic.  I am removing the test to create the
object due to the overhead of integrating with VCR for this simple test.
More complex tests will be added at a later date.