Add github pull request template
This adds a standard way of formatting a pull-request and guiding a contributer
to the prefered format and level of detail for a pull request.  This feature
was recently added to the github web UI:

https://help.github.com/articles/creating-a-pull-request-template-for-your-repository/

And is in the process of being supported in `hub`, the official command line
git augmentation tool for github:

https://github.com/github/hub/pull/1187

The idea is this is a suggestion, not an enforcement.  Everything in the quotes
is meant to be a short documentation of guidelines for the commiter to use when
opening a pull-request, and should be deleted and filled out with pull-request
context specific information.  None of the pull-request template is enforced by
github and can be deleted if the commiter so chooses.