MiqDisk Disk Cache Module
Add a new MiqDisk module which will do generic disk caching for other
MiqDisk modules.  To use it, instantiate the MiqDisk module in question and
pass it as an argument to the MiqDiskCache.new method.

Spec tests have been added for the object and its methods.

In addition a new dummy MiqDisk module (MiqDummyDisk) has been added to facilitate
the above spec tests (as well as more spec tests for the dummy disk module).