Cache storage.ems to make queue name lookup easier
For storage, we often don't need every ems, we just need the
first one we can find. Very often we used an ems to fetch the storage
in the first place.

In our usages, we often use an ems to fetch storages.
Resources that have a `belong_to` like `Vm` cache that value. So when
you ask for vm.ems, it will not need a database lookup.
Storages on the other hand, have many emses, so rails caching does not
work. It requires a query.

So for processes like cap&U, we end up with an extra query for every
storage. Not the end of the world but avoidable.