Trim down alert profile assignments call
This was calling N+1 queries to looking up all the tagging for every tag
that is assigned to an alert profile.
(Since the tags are specially formed for alert profiles this is not the
end of the world.)

Now, use a single query to bring back only the tagging of interest.

Since there is only 1 tag per object, the assignment_map is probably not
necessary, but left it as is.