rework assignments#assigned
store assigned as a hash of arrays
since these are looked up by key, makes more sense.

---

This may be overkill, storing as a hash of objects
may make more sense.

Alerts do not need an array. It just cares about presence
but chargebacks seem to want an array of objects