Remove the verify_local rake task
This is no longer useful as we may use a hostname to connect to
database servers that are actually running locally to "proxy"
connections so we can easily switch in a failover scenario.

Also we may run standby database servers when we are actually
configured against a remote primary server. In that case we
would want the local database to continue running even though
we are not configured to use it.