BAT Handling in Checkpoint Disks Issues
Address two problems handling the BAT for Checkpoint disks:

Depending upon the disk an attempt may be made to
access BAT information prior to initializing it, so initialize when appropriate.

Based on the VHDX spec the size of the BAT for checkpoint disks is:

Total BAT Entries = Sector Bitmap Blocks Count *  (Chunk Ratio + 1)

but there are checkpoint disks where the Sector Bitmap Blocks count is 0.
The size of the BAT for non-checkpoint disks is the following, and if the Sector
Bitmap Blocks Count is zero, we will use this formula.

Total BAT Entries = Data Blocks Count +  Floor((Data Blocks Count - 1) / Chunk Ratio)