relationship filter supports scopes
for cap&u run, it brings back 5% fewer rows but the real gain is
this query is kept as a scope. So it can be used as an inner query
and the total # queries, objects, are reduced.

This is just a building block

This slows the query down (less than margin of error).
We will make up for it by
reducing the total number of run