Adds scopes and helper methods to Classification
Adds some scopes to help prevent loading tags when fetching
classifications and also adds some convenience scopes.

tag_name is used to pulled the column from an instance of
`Classification`, if it exists, otherwise it returns nil.

When calling `name`, instead of fetching the associated tag, if the
tag_name column exists (when performing a join using the `with_tag_name`
or `managed` scope), it will use that, otherwise fetch the tag and role
with that.