Use a true stub for stub_user
This changes stub_user to use a non-persisted user object and changes
the :user_with_group factory to use a non-persisted group object. This
increases test performance as stub_user is called over 1,000 times in
controller specs, writing a user and group record to the database every
time. Automation specs also make heavy use of :user_with_group, so each
group write there is avoided.

As expected, stubbing is more performant but loses a little bit of true
parity with actual code testing. For example, because the user and group
are empty objects and have no primary keys, code that utilizes IDs
directly MUST use a real record. For that, one should use create an
actual user/group and use 'login_as'. Given that I needed to change only
a *single* spec because of this, it seems well worth it in 99% of cases
(the spec I have changed here uses current_user.miq_groups.ids in a
query for groups added to a report)

In short, stub_user vs. login_as should be used as most people are told
to use FactoryGirl's create vs. build - Use build as much as you can,
but create if you really need to.