Do not allow deletion of groups that are read-only
The read-only group is either system_group or tenant_group. The system
groups are hard to remove as rails will issue a ROLLBACK because of
cascade dependencies. The tenant_groups were introduced in 28005f8f and
we shall never allow deletion of them through api.

Also, prefer to raise Forbidden on read-only group edit, as we do the
same for categories.