rubocop issues: tree_builder_report_roles.rb
== app/presenters/tree_builder_report_roles.rb ==
W:  4: 25: Unused method argument - tree_name. If it's necessary, use _
or _tree_name as an argument name to indicate that it won't be used. You
can also write as tree_init_options(*) if you want the method to accept
any arguments but don't care about them.
C: 21:  5: Use the return of the conditional for variable assignment and
comparison.