Updating to newer version of net-ldap gem
Version 0.14.0 of the gem (actually, starting with 0.13.0) contains a code change that fixes an encoding error
(Encoding::UndefinedConversionError) that happens when there are extended characters in a dn. The fix forces
utf-8 encoding instead of ASCII-8BIT for objects returned from the directory.

See https://github.com/ruby-ldap/ruby-net-ldap/pull/242
https://bugzilla.redhat.com/show_bug.cgi?id=1367600