Add VMware cloud orchestration stack operations
This patch extends the orchestration stack of the VMware cloud manager
with the support for provisioning a new stack from a template and delete
an existing one. In VMware, the orchestration stack is represented by a
vApp, while it's template, stored in a catalog, is a vApp template.

The refresh parser is responsible for parsing these templates into
`orchestration_templates` table. However, due to the way vApps are
instantiated in VMware cloud, this patch does not use the content of the
template, but rather just uses it's ID and requests provisioning.

Corresponding tests are provided using mocks in a similar way as other
providers do it.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>