Add configuration for repmgrd to appliance_console
This will allow a user to start repmgrd for automatic failover
when configuring a standby server.

This requires us to write out a .pgpass file for the postgres user
so repmgrd can contact the primary database and also so it can
execute the proper commands when promoting the local server.

https://www.pivotaltracker.com/story/show/127384221