Report manufacturer and product info for RHEVM hosts
Save information about the operating system and product for
RHEVM hosts.

Previously, the product name used to be hard-coded 'linux' and no value
for the product type. With this patch the actual host os name as
reported by the provider will be stored for the host and its product
type will be set to 'linux'.

http://bugzilla.redhat.com/963905