Mapping CloudTenant to Tenant
Creation tree of Tenants according to CloudTenants, this tree is
attached to base tenant which is related to Openstack Provider.

CloudTenants and Tenants are tied thru relation Tenant#source and
on CloudTenant it is called CloudTenant#source_tenant

There is walking thru CloudTenant tree:

when association CloudTenant#source_tenant is found then
attributes of Tenant are updated.

when associations CloudTenant#source_tenant is not found then
Tenant is created. As parent is used CloudTenant#source_tenant
parent cloud_tenant and when such parent does not exist
it means that we are creating Tenant directly under base tenant.
So source_tenant of EmsCloud is in that case.