Remove faulty seed removal
You cannot use MiqExpression in migrations, because it loads application
code. This results in the good-migrations gem shutting you down.

The original PR (https://github.com/ManageIQ/manageiq/pull/9746) for
this just needs to remove a few MiqSearch records that are created
during the seeding process. The are unique and idempotent. There is no
reason that we shouldn't just look them up by name and remove them.

For the down half of the migration, we cannot easily recreate them as
they require serialized MiqExpression objects. Again, as these are
seeded values, they will be replaced should you be rolling back code
changes (ie, they'll be put back in fixtures and you'll be on your merry
way).

Because the migration is now completely trivial, testing only an up
migration, testing that the Rails API does it's job is pointless and
thus the entire spec file is removed.