Add AuthenticationMixin to MiqRegion model
This will be used to store remote region API authentication.

As a first pass, methods to fetch a remote region's v2_key
and to generate a token compatible with the the X-MIQ-Token
API authentication method added in
https://github.com/ManageIQ/manageiq/pull/10351 have been added.

The ultimate goal will be to use OAuth tokens in place of the
v2_key and contrived X-MIQ-Token authentication mechanism.

We store the entire remote region key text, encrypted using the
local region's v2_key, as an authentication record of type AuthToken
with the MiqRegion record as the resource.