Fix issues around specs replacing the global Settings object.
Prior to this change, MiqServer.seed would call Vmdb::Settings.init,
which would replace the global Settings object.  So, when this method
was called in specs, it would taint the global Settings object with one
that was attached to an MiqServer record that would eventually be
removed.  After that, any test that called Settings.reload! would fail
because that MiqServer record would not exist.

In addition, in specs, a user can create a "local" MiqServer object in a
number of ways, and it is expected that the Settings object will react
accordingly, so we can't just "ignore" the problem in specs.

This commit does 3 changes that work together.

1. Vmdb::Settings.init will no longer create a Settings object with the
  literal instance of MiqServer.my_server, but instead creates it with a
  placeholder that tells the DatabaseSource to dynamically load
  MiqServer.my_server on Settings.reload!.
2. Since Settings.reload! will dynamically fetch MiqServer.my_server,
  there is no need for MiqServer.seed to re-initialize the global
  Settings object, and can instead just call Settings.reload!
3. In specs, if the global Settings object has been reloaded, we will
  know, based on the value of Vmdb::Settings.last_loaded, so we can now
  undo any changes made by the tests.