Fix private/public class methods
All of these methods after the private are currently public.
If the intent was to make them private, only 2 of them could be private
since the others are called from outside this class on a receiver.

Therefore, use private_class_method on two of them, move the others to the public
area.

rubocop --only UselessAccessModifier