Relationship: skip unneeded is_isolated_root?
Don't call is_isloated_root? to short circuit the calls

The methods calls is_isolated_root? to ensure relationship is not null
but the very next call it looks up the relationship. Best to just fetch
the relationship and then react accordingly.

This short circuit often makes an additional query. Sometimes the code
could have chained query through to rbac, but instead it forces a
load and database queries.

A cache was needed to alleviate the extra is_isolated_root? query.

Also, the short circuit is skipping the filtering in the _rel_ methods.